requirements.txt
appdirs==1.4.4
certifi==2020.12.5
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
flask-download-btn==0.0.22
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
opencv-contrib-python==4.5.1.48
Pillow==8.1.2
pipenv==2020.11.15
pytesseract==0.3.7
six==1.15.0
SQLAlchemy==1.4.2
sqlalchemy-modelid==0.0.3
sqlalchemy-mutable==0.0.13
virtualenv==20.4.3
virtualenv-clone==0.5.4
Werkzeug==1.0.1
